DEV-1449: schema updates for holdings/hathifile loading

+++ sql/005_cluster_ocns.sql
@@ -2,8 +2,8 @@ USE `ht_repository`;
 
 DROP TABLE IF EXISTS `cluster_ocns`;
 CREATE TABLE `cluster_ocns` (
-  cluster_id int NOT NULL,
-  ocn int NOT NULL,
+  cluster_id int unsigned NOT NULL,
+  ocn bigint unsigned NOT NULL,
   PRIMARY KEY (ocn),
   INDEX (cluster_id)
 ) ENGINE=InnoDB;
